requirements.txt
ccxt>=4.4.80
pandas>=2.2.3
numpy>=2.1.3
python-dotenv>=1.0.1
python-telegram-bot>=21.10
tenacity>=9.0.0
tzdata>=2024.2
ta>=0.11.0
